Assignment: Create a Macro Script II
Similar to the previous exercise, use MEL to create an additional model. For this exercise, utilize variables and data types to store and retrieve data. This practice is imperative for isolating and targeting objects in Maya commands.

The completed model must meet the following requirements:

All geometry is  combined into a single object
The model is renamed appropriately
The model is parented under a top-level "Geometry" group
At a minimum, execute the following list of actions using commands and parameters:

Commands for Polygon Primitives
Extrusions
Move, Rotate, and Scale
Parent
Group
Rename